refactor: useSearch 훅 분리
